requirements.txt
boto3==1.6.20
botocore==1.9.20
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.3
django-bootstrap-form==3.4
django-isbn-field==0.5.1
django-storages==1.6.6
docutils==0.14
gunicorn==19.7.1
jmespath==0.9.3
olefile==0.45.1
Pillow==5.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.6.1
python-decouple==3.1
python-stdnum==1.8.1
pytz==2018.3
s3transfer==0.1.13
six==1.11.0
sqlparse==0.2.4
static3==0.7.0
Unipath==1.1
whitenoise==3.3.1